requirement.txt
einops==0.8.0
flash_attn==2.5.8
kornia==0.7.2
omegaconf==2.3.0
open_clip_torch
opencv_python
pytorch_lightning==2.2.4
safetensors==0.4.3
scipy
SwissArmyTransformer
torch
torchvision
transformers
webdataset